requirements.txt
Flask
Flask-SQLAlchemy
Flask-Restful
Flask-WTF
Flask-Migrate
Flask-marshmallow
Flask-restful
Flask-Pydantic
python-jose[cryptography]
passlib[bcrypt]
psycopg2
marshmallow-sqlalchemy
pydantic[email]
flask-jwt-extended